backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
reportlab==4.0.7
openpyxl==3.1.2
XlsxWriter==3.1.9
google-generativeai==0.3.1
Pillow==10.1.0
python-dotenv==1.0.0
